+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-checkbox-multiple/src/main/java/com/liferay/dynamic/data/mapping/type/checkbox/multiple/internal/CheckboxMultipleDDMFormFieldRenderer.java
@@ -57,7 +57,7 @@ public class CheckboxMultipleDDMFormFieldRenderer
 	@Activate
 	protected void activate(Map<String, Object> properties) {
 		_templateResource = getTemplateResource(
-			"/META-INF/resources/checkbox_multiple.soy");
+			"/META-INF/resources/checkbox-multiple.soy");
 	}
 
 	@Deactivate
