+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-checkbox-multiple/src/main/resources/META-INF/resources/checkbox-multiple.es.js
@@ -1,7 +1,7 @@
 import Component from 'metal-component';
 import Soy from 'metal-soy';
 
-import templates from './checkbox_multiple.soy';
+import templates from './checkbox-multiple.soy';
 
 /**
  * CheckboxMultiple Component
