+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-checkbox-multiple/src/main/resources/META-INF/resources/checkbox-multiple.js
@@ -1,6 +1,6 @@
 Liferay.Loader.require(
 	'frontend-js-metal-web/metal-soy-bundle/build/bundle',
-	'dynamic-data-mapping-type-checkbox-multiple/checkbox_multiple.es',
+	'dynamic-data-mapping-type-checkbox-multiple/checkbox-multiple.es',
 	function(MetalSoyBundle, CheckboxMultiple) {
 		if (!window.DDMCheckboxMultiple) {
 			window.DDMCheckboxMultiple = {};
