+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-checkbox-multiple/src/main/resources/META-INF/resources/checkbox-multiple.soy
