+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-checkbox-multiple/src/main/resources/META-INF/resources/config.js
@@ -33,7 +33,7 @@
 							condition: {
 								trigger: 'liferay-ddm-form-renderer'
 							},
-							path: 'checkbox_multiple.js'
+							path: 'checkbox-multiple.js'
 						}
 					},
 					root: MODULE_PATH + '/'
