+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-key-value/src/main/java/com/liferay/dynamic/data/mapping/type/key/value/internal/KeyValueDDMFormFieldRenderer.java
@@ -56,7 +56,7 @@ public class KeyValueDDMFormFieldRenderer extends BaseDDMFormFieldRenderer {
 	@Activate
 	protected void activate(Map<String, Object> properties) {
 		_templateResource = getTemplateResource(
-			"/META-INF/resources/key_value.soy");
+			"/META-INF/resources/key-value.soy");
 	}
 
 	@Deactivate
