+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-key-value/src/main/resources/META-INF/resources/config.js
@@ -24,7 +24,7 @@
 							condition: {
 								trigger: 'liferay-ddm-form-renderer'
 							},
-							path: 'key_value.js'
+							path: 'key-value.js'
 						}
 					},
 					root: MODULE_PATH + '/'
