+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-key-value/src/main/resources/META-INF/resources/key-value.es.js
@@ -1,7 +1,7 @@
 import Component from 'metal-component';
 import Soy from 'metal-soy';
 
-import templates from './key_value.soy';
+import templates from './key-value.soy';
 
 /**
  * KeyValue Component
