+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-key-value/src/main/resources/META-INF/resources/key-value.soy
